Fix implicit conversion warning

+++ mifare.c
@@ -59,7 +59,7 @@ nfc_initiator_mifare_cmd(nfc_device *pnd, const mifare_cmd mc, const uint8_t ui8
   uint8_t  abtCmd[265];
   //bool    bEasyFraming;
 
-  abtCmd[0] = mc;               // The MIFARE Classic command
+  abtCmd[0] = (uint8_t)mc;      // The MIFARE Classic command
   abtCmd[1] = ui8Block;         // The block address (1K=0x00..0x39, 4K=0x00..0xff)
 
   switch (mc) {
